update: Fixing workflow
